requirements.txt
pandas>=1.5.0
numpy>=1.23.0
matplotlib>=3.7.0
plotly>=5.20.0
openpyxl>=3.1.0
requests>=2.28.0
eth-account>=0.8.0

# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0

# Authentication
PyJWT>=2.8.0
passlib>=1.7.4
argon2-cffi>=23.1.0

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0
